Fast fix for test.

+++ MangaReader.Core/Update/Updater.cs
@@ -55,7 +55,7 @@ namespace MangaReader.Core.Update
       try
       {
         var x64folder = Path.Combine(ConfigStorage.WorkFolder, "x64");
-        if (Directory.Exists(x64folder))
+        if (Directory.Exists(x64folder) && System.Environment.GetCommandLineArgs().All(a => !a.Contains("vstest")))
         {
           File.Delete(Path.Combine(ConfigStorage.WorkFolder, "sqlite3"));
           File.Delete(Path.Combine(ConfigStorage.WorkFolder, "System.Data.SQLite.dll"));
